backend/requirements.txt
fastapi
uvicorn[standard]
python-socketio
langgraph
langchain
langchain-ollama
langchain-community
pdfplumber
httpx
python-dotenv
aiosqlite
anthropic
